requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
Django==5.2
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
phonenumbers==9.0.2
psycopg==3.2.6
psycopg2==2.9.10
pycparser==2.22
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.1.0
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.13.1
uvicorn==0.34.0
whitenoise==6.9.0
